requirements.txt
telethon
tweepy
python-dotenv
aiofiles
pillow
colorlog